removed interactive promp from pype.py file. renamed method in vmtk.run
