+++ __init__.py
@@ -1,2 +1,5 @@
-from . import vmtkscripts
-from . import pypes
+from .vmtkscripts import *
+from .pypes import *
+from .interactivepype import *
+
+__all__ = [interactivepype.__all__]
